+++ src/main/java/kitchenpos/repository/MenuProductRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.repository;
+
+import kitchenpos.domain.MenuProduct;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
+    List<MenuProduct> findAllByMenuId(final Long menuId);
+}
