+++ src/main/java/kitchenpos/repository/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.repository;
 
 import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MenuRepository extends JpaRepository<Menu,Long> {
-    long countByIdIn(List<Long> ids);
+    long countByIdIn(final List<Long> ids);
 }
