+++ src/main/java/kitchenpos/repository/OrderLineItemRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.repository;
 
 import kitchenpos.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem,Long> {
-    List<OrderLineItem> findAllByOrderId(Long orderId);
+    List<OrderLineItem> findAllByOrderId(final Long orderId);
 }
